* 'master' of git://github.com/thelia/thelia: (126 commits)
  fix module template inclusion
  fix modules bug
  configure button in module page
  fix class name in module generation
  fix unit tests search by id in loop bug due to not visible first entry in database
  colissimo
  colissimo module
  colissimo module
  Colissimo in progress
  Colissimo proces
  Colissimo module + countires areas
  Fixed typos
  change translation template baseDir
  fix typo
  check extensions in install process
  module template inclusion
  change template subdirs
  admin for colissimo
  module edition
  ISSUE 100 : Information has no plural (Fix front)
  ...

+++ core/lib/Thelia/Action/Customer.php
@@ -60,7 +60,7 @@ class Customer extends BaseAction implements EventSubscriberInterface
 
     }
 
-    public function updateProfil(CustomerCreateOrUpdateEvent $event)
+    public function updateProfile(CustomerCreateOrUpdateEvent $event)
     {
 
         $customer = $event->getCustomer();
@@ -83,9 +83,10 @@ class Customer extends BaseAction implements EventSubscriberInterface
 
     public function delete(CustomerEvent $event)
     {
-        $customer = $event->getCustomer();
+        if (null !== $customer = $event->getCustomer()) {
 
-        $customer->delete();
+            $customer->delete();
+        }
     }
 
     private function createOrUpdateCustomer(CustomerModel $customer, CustomerCreateOrUpdateEvent $event)
@@ -166,7 +167,7 @@ class Customer extends BaseAction implements EventSubscriberInterface
         return array(
             TheliaEvents::CUSTOMER_CREATEACCOUNT    => array('create', 128),
             TheliaEvents::CUSTOMER_UPDATEACCOUNT    => array('modify', 128),
-            TheliaEvents::CUSTOMER_UPDATEPROFIL     => array('updateProfil', 128),
+            TheliaEvents::CUSTOMER_UPDATEPROFILE     => array('updateProfile', 128),
             TheliaEvents::CUSTOMER_LOGOUT           => array('logout', 128),
             TheliaEvents::CUSTOMER_LOGIN            => array('login', 128),
             TheliaEvents::CUSTOMER_DELETEACCOUNT    => array('delete', 128),
